requirements.txt
pyproj
moderngl
Shapely
#owslib
PILLOW
diskcache
pyrr
moderngl_window #TODO generify